Better typing on Arrays.
Added typing to array of palettes.  This makes understanding which palette you're working with Bright vs. Normal.  
Moved the creation of the palette to process_ansi method. This method is the only method that uses palette_256. Why pay the price for code that may never be executed. There are several public methods available on your class that dont utilize this object.