Fix handling of faint-only style
This is a fix of the PR done in #93. That PR seems to have updated the generated ansi_up.js file instead of ansi_up.ts, which caused the file to be overwritten when the version bump recompiled the files in 65baddb93b16e75cf5a33104501dd1cfa921f404. This PR performs the same change as #93, but in ansi_up.ts instead of ansi_up.js.